% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALC.R
\name{ALC_two_level}
\alias{ALC_two_level}
\title{find the next point by ALC criterion with two fidelity levels and update the model}
\usage{
ALC_two_level(
  Xref = NULL,
  fit,
  mc.sample = 100,
  cost,
  funcs,
  n.start,
  parallel = FALSE,
  ncore = 1
)
}
\arguments{
\item{Xref}{vector or matrix of reference data.}

\item{fit}{an object of class RNAmf.}

\item{mc.sample}{a number of mc samples generated for this approach. Default is 100.}

\item{cost}{a vector of the costs for each level of fidelity.}

\item{funcs}{list of functions for each level of fidelity.}

\item{n.start}{a number of candidate point. Default is 10*d.}

\item{parallel}{logical indicating whether to run parallel or not. Default is FALSE.}

\item{ncore}{the number of core for parallel. Default is 1.}
}
\value{
A list containing the integrated one-step ahead variance:
\itemize{
  \item \code{fit}: fitted model after acquire chosen point.
  \item \code{intvar1}: vector of integrated variance when each data point is added at level 1.
  \item \code{intvar2}: vector of integrated variance when each data point is added at level 2.
  \item \code{cost}: vector of the costs for each level of fidelity.
  \item \code{Xcand}: candidate data points considered to be acquired.
  \item \code{chosen}: list of chosen level, location, and point.
}
}
\description{
The function acquires the new point in two fidelity levels setting
by the Active learning Cohn(ALC) criterion.
The function generates the candidate set by maximin design,
and compute the reduction in posterior variances across the entire input space after running the selected point and level.
The outputs to calculate the reduction are imputed by MC approximation.
Then, it optimizes the ALC criterion starting at the candidate point with maximizes ALC criterion.
After optimization, the function calculates the ratio between the reduction in posterior variances and the simulation cost.
Finally, it acquires the point and the level maximizing the ALC criterion.
}
