% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALC_two_level_optm.R
\name{ALC_two_level_optm}
\alias{ALC_two_level_optm}
\title{ALC_two_level_optm}
\usage{
ALC_two_level_optm(Xref, fit, cost, funcs)
}
\arguments{
\item{Xref}{vector or matrix of reference data.}

\item{fit}{an object of class RNAmf.}

\item{cost}{a vector of the costs for each level of fidelity.}

\item{funcs}{list of functions for each level of fidelity.}
}
\value{
A list containing the integrated one-step ahead variance:
\itemize{
  \item \code{fit}: fitted model after acquire chosen point.
  \item \code{intvar1}: vector of integrated variance when each data point is added at level 1.
  \item \code{intvar2}: vector of integrated variance when each data point is added at level 2.
  \item \code{cost}: vector of the costs for each level of fidelity.
  \item \code{Xcand}: candidate data points considered to be acquired.
  \item \code{chosen}: list of chosen level, location, and point.
}
}
\description{
find the next point by ALC criterion with two fidelity levels and update the model
}
