% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALM.R
\name{ALM_three_level}
\alias{ALM_three_level}
\title{find the next point by ALM criterion with three fidelity levels and update the model}
\usage{
ALM_three_level(fit, cost, funcs, n.start, parallel = FALSE, ncore = 1)
}
\arguments{
\item{fit}{an object of class RNAmf.}

\item{cost}{a vector of the costs for each level of fidelity.}

\item{funcs}{list of functions for each level of fidelity.}

\item{n.start}{a number of candidate point. Default is 10*d.}

\item{parallel}{logical indicating whether to run parallel or not. Default is FALSE.}

\item{ncore}{the number of core for parallel.}
}
\value{
A list containing the integrated one-step ahead variance:
\itemize{
  \item \code{fit}: fitted model after acquire chosen point.
  \item \code{predsig2.low}: vector of predictive variance of optimized data point at level 1.
  \item \code{predsig2.med}: vector of predictive variance of optimized data point at level 2.
  \item \code{predsig2.high}: vector of predictive variance of optimized data point at level 3.
  \item \code{cost}: vector of the costs for each level of fidelity.
  \item \code{Xcand}: candidate data points considered to be acquired.
  \item \code{chosen}: list of chosen level, location, and point.
}
}
\description{
The function acquires the new point in three fidelity levels setting
by the Active learning MacKay(ALM) criterion.
The function generates the candidate set by maximin design,
and compute the posterior predictive variance at each level.
Then, it optimizes the ALM criterion starting at the candidate point with the maximum posterior predictive variance.
After optimization, the function calculates the ratio between the posterior predictive variance and the simulation cost.
Finally, it acquires the point and the level maximizing the ALM criterion.
}
