% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closed.R
\name{closed}
\alias{closed}
\title{closed}
\usage{
closed(X1, y1, X2, y2, kernel, constant = FALSE)
}
\arguments{
\item{X1}{vector or matrix of input locations for the low fidelity level.}

\item{y1}{vector of response values for the low fidelity level.}

\item{X2}{vector or matrix of input locations for the high fidelity level.}

\item{y2}{vector of response values for the high fidelity level.}

\item{kernel}{chracter specifying kernel type to be used, to be chosen between "sqex"(squared exponential), "matern1.5", or "matern2.5"}

\item{constant}{logical indicating for constant mean (constant=TRUE) or zero mean (constant=FALSE). Default is FALSE.}
}
\value{
A list containing fitted models f1 and f2, constant, and kernel structure:
\itemize{
  \item \code{fit1}: list of fitted model for (X1, y1).
  \item \code{fit2}: list of fitted model for ((X2, f1(X2)), y2).
  \item \code{constant}: copy of constant.
  \item \code{kernel}: copy of kernel.
}
}
\description{
Fitting the model with two fidelity levels
}
