% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor.sep.R
\name{cor.sep}
\alias{cor.sep}
\title{cor.sep}
\usage{
cor.sep(X, x = NULL, theta, nu, derivative = 0)
}
\arguments{
\item{X}{vector or matrix of input.}

\item{x}{vector or matrix of new input. Default is NULL}

\item{theta}{lengthscale parameter. It should have the length of ncol(X).}

\item{nu}{numerical value of smoothness hyperparameter. It should be 0.5, 1.5, 2.5, 3.5, or 4.5.}

\item{derivative}{logical indicating for its first derivative(derivative=1)}
}
\value{
A covariance matrix of matern kernel.
}
\description{
calculating separable matern kernel
}
