% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GP.R
\name{GP}
\alias{GP}
\title{GP}
\usage{
GP(
  X,
  y,
  g = sqrt(.Machine$double.eps),
  lower = 0.001,
  upper = 1000,
  Xscale = TRUE,
  Yscale = TRUE,
  constant = FALSE
)
}
\arguments{
\item{X}{vector or matrix of input locations.}

\item{y}{vector of response values.}

\item{g}{nugget parameter. Default is 1.490116e-08.}

\item{lower}{lower bound of theta. Default if 0.001.}

\item{upper}{upper bound of theta. Default if 1000.}

\item{Xscale}{logical indicating whether to scale X or not. Default is TRUE.}

\item{Yscale}{logical indicating whether to scale y or not. Only used if constant=FALSE. Default is TRUE.}

\item{constant}{logical indicating for constant mean (constant=TRUE) or zero mean (constant=FALSE). Default is FALSE.}
}
\value{
A list containing hyperparameters, covariance inverse matrix, X, y and logical inputs:
\itemize{
  \item \code{theta}: vector of lengthscale hyperparameter.
  \item \code{g}: copy of g.
  \item \code{Ki}: matrix of covariance inverse.
  \item \code{mu.hat}: optimized constant mean. If constant=FALSE, 0.
  \item \code{X}: copy of X. If Xscale=TRUE, scaled X.
  \item \code{y}: copy of y. If Yscale=TRUE, scaled y.
  \item \code{tau2hat}: estimated scale hyperparameter.
  \item \code{Xscale}: copy of Xscale.
  \item \code{Yscale}: copy of Yscale.
  \item \code{constant}: copy of constant.
}
}
\description{
fitting the model with squared exponential kernel.
}
\examples{
\dontrun{
library(lhs)
### synthetic function ###
f1 <- function(x)
{
  sin(8*pi*x)
}

### training data ###
n1 <- 15

X1 <- maximinLHS(n1, 1)
y1 <- f1(X1)

GP(X1, y1)
}
}
