% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integvar.R
\name{integvar}
\alias{integvar}
\title{integvar}
\usage{
integvar(x, fit, mc.sample = 10)
}
\arguments{
\item{x}{vector or matrix of test data.}

\item{fit}{an object of class RNAmf.}

\item{mc.sample}{a number of mc samples generated for this approach. Default is 10.}
}
\value{
A list containing the integrated one-step ahead variance:
\itemize{
  \item \code{intvar1}: vector of integrated variance when each data point is added at level 1.
  \item \code{intvar2}: vector of integrated variance when each data point is added at level 2.
}
}
\description{
computing integrated one-step ahead variance with two fidelity levels
}
