% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integvar2.R
\name{integvar2}
\alias{integvar2}
\title{integvar2}
\usage{
integvar2(x, fit, mc.sample = 10000)
}
\arguments{
\item{x}{vector or matrix of test data.}

\item{fit}{an object of class closed2.}

\item{mc.sample}{a number of mc samples generated for this approach. Default is 10000.}
}
\value{
A list containing the integrated one-step ahead variance:
\itemize{
  \item \code{intvar1}: vector of integrated variance when each data point is added at level 1.
  \item \code{intvar2}: vector of integrated variance when each data point is added at level 2.
  \item \code{intvar3}: vector of integrated variance when each data point is added at level 3.
}
}
\description{
computing integrated one-step ahead variance with three fidelity levels
}
