% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matern.kernel.R
\name{matern.kernel}
\alias{matern.kernel}
\title{matern.kernel}
\usage{
matern.kernel(r, nu, derivative = 0)
}
\arguments{
\item{r}{vector or matrix of input.}

\item{nu}{numerical value of smoothness hyperparameter. It should be 0.5, 1.5, 2.5, 3.5, or 4.5.}

\item{derivative}{logical indicating for its first derivative(derivative=1)}
}
\value{
A value from matern kernel.
}
\description{
calculating matern kernel with corresponding smoothness parameter
}
