% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALM_two_level.R
\name{obj.ALM_two_level_2}
\alias{obj.ALM_two_level_2}
\title{object to optimize the point by ALM criterion updating at level 2 with two levels of fidelity}
\usage{
obj.ALM_two_level_2(Xcand, fit)
}
\arguments{
\item{Xcand}{candidate data point to be optimized.}

\item{fit}{an object of class RNAmf.}
}
\value{
A mean of the negative predictive posterior variance at Xcand.
}
\description{
object to optimize the point by ALM criterion updating at level 2 with two levels of fidelity
}
