% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred.matGP.R
\name{pred.matGP}
\alias{pred.matGP}
\title{pred.matGP}
\usage{
pred.matGP(fit, xnew)
}
\arguments{
\item{fit}{an object of class matGP.}

\item{xnew}{vector or matrix of new input locations to predict.}
}
\value{
A list predictive posterior mean and variance:
\itemize{
  \item \code{mu}: vector of predictive posterior mean.
  \item \code{sig2}: vector of predictive posterior variance.
}
}
\description{
predictive posterior mean and variance with matern kernel.
}
\examples{
\dontrun{
library(lhs)
### synthetic function ###
f1 <- function(x)
{
  sin(8*pi*x)
}

### training data ###
n1 <- 15

X1 <- maximinLHS(n1, 1)
y1 <- f1(X1)

fit1 <- matGP(X1, y1, nu=2.5)

### test data ###
x <- seq(0,1,0.01)
pred.matGP(fit1, x)
}
}
