% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predRNAM2.R
\name{predRNAM2}
\alias{predRNAM2}
\title{predRNAM2}
\usage{
predRNAM2(fit, x)
}
\arguments{
\item{fit}{an object of class RNAM2.}

\item{x}{vector or matrix of new input locations to predict.}
}
\value{
A list predictive posterior mean and variance:
\itemize{
  \item \code{mu}: vector of predictive posterior mean.
  \item \code{sig2}: vector of predictive posterior variance.
}

#' @importFrom plgp distance
}
\description{
predictive posterior mean and variance of the model with fidelity level 1, 2 and 3.
}
