% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predRNAmf.R
\name{predRNAmf}
\alias{predRNAmf}
\title{predRNAmf}
\usage{
predRNAmf(fit, x)
}
\arguments{
\item{fit}{an object of class RNAmf.}

\item{x}{vector or matrix of new input locations to predict.}
}
\value{
A list predictive posterior mean and variance:
\itemize{
  \item \code{mu}: vector of predictive posterior mean.
  \item \code{sig2}: vector of predictive posterior variance.
}

#' @importFrom plgp distance
}
\description{
predictive posterior mean and variance of the model with fidelity level 1 and 2.
}
\examples{
library(MuFiCokriging)
library(lhs)

### synthetic function ###
f1 <- function(x)
{
  sin(8*pi*x)
}

f2 <- function(x)
{
  (x-sqrt(2))*(sin(8*pi*x))^2
}

### training data ###
n1 <- 15; n2 <- 10

X1 <- maximinLHS(n1, 1)
X2 <- maximinLHS(n2, 1)

NestDesign <- NestedDesignBuild(design = list(X1,X2))

X1 <- NestDesign$PX
X2 <- ExtractNestDesign(NestDesign,2)

y1 <- f1(X1)
y2 <- f2(X2)

### test data ###
x <- seq(0,1,length.out=100)

### fitting ###
fit.RNAmf <- RNAmf(X1, y1, X2, y2, kernel="sqex", constant=TRUE)
predy <- predRNAmf(fit.RNAmf, x)$mu
predsig2 <- predRNAmf(fit.RNAmf, x)$sig2

}
