% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predRNAmf_two_level.R
\name{predRNAmf_two_level}
\alias{predRNAmf_two_level}
\title{predictive posterior mean and variance of the model with fidelity level 1 and 2.}
\usage{
predRNAmf_two_level(fit, x)
}
\arguments{
\item{fit}{an object of class RNAmf.}

\item{x}{vector or matrix of new input locations to predict.}
}
\value{
A list predictive posterior mean and variance:
\itemize{
  \item \code{mu}: vector of predictive posterior mean.
  \item \code{sig2}: vector of predictive posterior variance.
}

#' @importFrom plgp distance
}
\description{
The function computes the posterior mean and variance of RNA models with two fidelity levels
by fitted model using \code{\link{RNAmf_two_level}}.
}
\details{
From the model fitted by \code{\link{RNAmf_two_level}},
the posterior mean and variance are calculated based on the closed form expression derived by a recursive fashion.
The formulae depend on its kernel choices.
For Matern kernels, \code{\link{zetafun}} computes the part of the posterior variance.
For details, see Heo and Sung (2023+, <arXiv:2309.11772>).
}
\examples{
library(MuFiCokriging)
library(lhs)

### Perdikaris function ###
f1 <- function(x)
{
  sin(8*pi*x)
}

f2 <- function(x)
{
  (x-sqrt(2))*(sin(8*pi*x))^2
}

### training data ###
n1 <- 13; n2 <- 8

X1 <- maximinLHS(n1, 1)
X2 <- maximinLHS(n2, 1)

NestDesign <- NestedDesignBuild(design = list(X1,X2))

X1 <- NestDesign$PX
X2 <- ExtractNestDesign(NestDesign,2)

y1 <- f1(X1)
y2 <- f2(X2)

### test data ###
x <- seq(0,1,length.out=1000)

### fitting ###
fit.RNAmf <- RNAmf_two_level(X1, y1, X2, y2, kernel="sqex", constant=TRUE)
pred.RNAmf <- predRNAmf_two_level(fit.RNAmf, x)
predy <- pred.RNAmf$mu
predsig2 <- pred.RNAmf$sig2

### RMSE ###
sqrt(mean((predy-f2(x))^2))

}
\seealso{
\code{\link{RNAmf_two_level}} for prediction.
}
