% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAM.R
\name{RNAM}
\alias{RNAM}
\title{RNAM}
\usage{
RNAM(X1, y1, X2, y2, kernel, constant = FALSE)
}
\arguments{
\item{X1}{vector or matrix of input locations for the low fidelity level.}

\item{y1}{vector of response values for the low fidelity level.}

\item{X2}{vector or matrix of input locations for the high fidelity level.}

\item{y2}{vector of response values for the high fidelity level.}

\item{kernel}{chracter specifying kernel type to be used, to be chosen between "sqex"(squared exponential), "matern1.5", or "matern2.5"}

\item{constant}{logical indicating for constant mean (constant=TRUE) or zero mean (constant=FALSE). Default is FALSE.}
}
\value{
A list containing fitted models f1 and f2, constant, and kernel structure:
\itemize{
  \item \code{fit1}: list of fitted model for (X1, y1).
  \item \code{fit2}: list of fitted model for ((X2, f1(X2)), y2).
  \item \code{constant}: copy of constant.
  \item \code{kernel}: copy of kernel.
}
}
\description{
Fitting the model with two fidelity levels
}
\examples{
library(MuFiCokriging)
library(lhs)

### synthetic function ###
f1 <- function(x)
{
  sin(8*pi*x)
}

f2 <- function(x)
{
  (x-sqrt(2))*(sin(8*pi*x))^2
}

### training data ###
n1 <- 15; n2 <- 10

X1 <- maximinLHS(n1, 1)
X2 <- maximinLHS(n2, 1)

NestDesign <- NestedDesignBuild(design = list(X1,X2))

X1 <- NestDesign$PX
X2 <- ExtractNestDesign(NestDesign,2)

y1 <- f1(X1)
y2 <- f2(X2)

### fitting ###
fit.RNAM <- RNAM(X1, y1, X2, y2, kernel="sqex", constant=TRUE)

}
