% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAM2.R
\name{RNAM2}
\alias{RNAM2}
\title{RNAM2}
\usage{
RNAM2(X1, y1, X2, y2, X3, y3, kernel, constant = FALSE)
}
\arguments{
\item{X1}{vector or matrix of input locations for the low fidelity level.}

\item{y1}{vector of response values for the low fidelity level.}

\item{X2}{vector or matrix of input locations for the medium fidelity level.}

\item{y2}{vector of response values for the medium fidelity level.}

\item{X3}{vector or matrix of input locations for the high fidelity level.}

\item{y3}{vector of response values for the high fidelity level.}

\item{kernel}{chracter specifying kernel type to be used, to be chosen between "sqex"(squared exponential), "matern1.5", or "matern2.5"}

\item{constant}{logical indicating for constant mean (constant=TRUE) or zero mean (constant=FALSE). Default is FALSE.}
}
\value{
A list containing fitted models f1, f2 and f3, constant, and kernel structure:
\itemize{
  \item \code{fit.RNAM1}: list of fitted model for low fidelity level (X1, y1) and medium fidelity level ((X2, f1(X2)), y2).
  \item \code{fit3}: list of fitted model for ((X2, f2(X3, f1(X3))), y3).
  \item \code{constant}: copy of constant.
  \item \code{kernel}: copy of kernel.
}
}
\description{
Fitting the model with three fidelity levels
}
