% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zetafun.R
\name{zetafun}
\alias{zetafun}
\title{computing zeta component in the closed form posterior variance of matern kernel.}
\usage{
zetafun(w1, w2, m, s, nu, theta)
}
\arguments{
\item{w1}{numerical value of f_1(x_k).}

\item{w2}{numerical value of f_1(x_l).}

\item{m}{numerical value of mu^*_1(x).}

\item{s}{numerical value of sigma^*2_1(x).}

\item{nu}{numerical value of smoothness hyperparameter. It should be 1.5 or 2.5.}

\item{theta}{vector of lengthscale hyperparameter.}
}
\value{
calculated value of zeta component.
}
\description{
computing zeta component in the closed form posterior variance of matern kernel.
}
