% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALC_two_level.R
\name{obj.ALC_two_level_1}
\alias{obj.ALC_two_level_1}
\title{object to optimize the next point by ALC criterion updating at level 1 with two levels of fidelity}
\usage{
obj.ALC_two_level_1(Xcand, Xref, fit, mc.sample, parallel = FALSE, ncore = 1)
}
\arguments{
\item{Xcand}{candidate data point to be optimized.}

\item{Xref}{vector or matrix of reference data.}

\item{fit}{an object of class RNAmf.}

\item{mc.sample}{a number of mc samples generated for this approach. Default is 100.}

\item{parallel}{logical indicating whether to run parallel or not. Default is FALSE.}

\item{ncore}{the number of core for parallel. Default is 1.}
}
\value{
A mean of the deduced variance at Xref.
}
\description{
object to optimize the next point by ALC criterion updating at level 1 with two levels of fidelity
}
