% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALC_two_level_optm.R
\name{obj.ALC_two_level_2}
\alias{obj.ALC_two_level_2}
\title{obj.ALC_two_level_2}
\usage{
obj.ALC_two_level_2(Xcand, Xref, fit)
}
\arguments{
\item{Xcand}{candidate data point to be optimized.}

\item{Xref}{vector or matrix of reference data.}

\item{fit}{an object of class RNAmf.}
}
\value{
A mean of the predictive posterior variance at Xref.
}
\description{
object to optimize the next point by ALC criterion updating at level 2 with two levels of fidelity
}
